fix installation issues
