+++ primitives/python/pulser_qrmi_backend/VERSION.txt
