+++ primitives/python/pulser_qrmi_backend/examples/pasqal/requirements.txt
