+++ primitives/python/qiskit_qrmi_primitives/LICENSE.txt
@@ -1,4 +1,4 @@
-                    Copyright 2024 IBM and its contributors
+                    Copyright 2025 IBM and its contributors
 
                                  Apache License
                            Version 2.0, January 2004
